+++ builder_pkgs/mockito/lib/src/mock.dart
@@ -235,7 +235,7 @@ class SmartFake {
   // make Analyzer happy, if we fake classes that override `==` to
   // accept `Object?` or `dynamic` (most notably [Interceptor]).
   @override
-  // ignore: non_nullable_equals_parameter
+  // ignore: non_nullable_equals_parameter, hash_and_equals
   bool operator ==(Object? other) => identical(this, other);
 
   @override